添加了一些结构体和变量准备添加seek功能

+++ YtxPlayer/app/src/main/java/com/ytx/ican/media/player/pragma/IMediaPlayer.java
@@ -82,7 +82,7 @@ public interface IMediaPlayer {
 
     boolean isPlaying();
 
-    void seekTo(long msec) throws IllegalStateException;
+    void seekTo(int msec) throws IllegalStateException;
 
     long getCurrentPosition();
